Use unsigned long for uploaded value

+++ src/TorrentRatioBooster/Services/UrlModifierService.cs
@@ -27,7 +27,7 @@ namespace TorrentRatioBooster.Services
             }
 
             var downloadedValue = queryStringList.FirstOrDefault(x => x.Key.Equals("downloaded", StringComparison.OrdinalIgnoreCase)).Value;
-            var uploadedValue = Convert.ToInt64(Math.Ceiling(Convert.ToInt64(downloadedValue) * ratio));
+            var uploadedValue = (ulong)Convert.ToInt64(Math.Ceiling(Convert.ToUInt64(downloadedValue) * ratio));
 
             this.logger.LogInformation($"Reporting downloaded: {downloadedValue} - Reporting uploaded: {uploadedValue}");
